core/skills/deepcamera/skills/automation/mqtt/requirements.txt
# MQTT Automation
paho-mqtt>=2.0.0
